pyproject.toml
[project]
name = "contactform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.177.0",
    "playwright>=1.54.0",
    "playwright-stealth>=2.0.0",
    "python-dotenv>=1.1.1",
]
